Add Swift 3/4 'servald' module and executable
If the Swift 3 or Swift 4 compiler is present or passed to the configure
script in the SWIFTC variable, then the Makefile will create a Swift
module called 'servald' for the Serval daemon and the 'servaldswift'
executable to ensure that the module links correctly into a stand-alone
Swift program.

Uses the gold linker if supported, to avoid relocation errors on symbols
produced by Swift when linking dynamic libraries.