Add Swift Keyring API with tests
The configure script optionally accepts the SWIFT_TARGET variable to
specify a build target for native or cross-compilation, and ensures that
the target is >= macosx10.10 when compiling for Mac OS-X, so that the
package manager will succeed.

Add autoconf macros for the Swift package manager.

Add a swift-api subdirectory containing the Swift source code and a
Makefile.in that compiles it into the "ServalClient" Swift package using
the Swift package manager.  The Swift API contains the following classes:

- ServalKeyring provides the operations: add, remove, set, list
- AbstractId and its specialisation SubscriberId, already in near-final
  form, are data types for SID and the like
- ServalRestfulClient (internal) uses an HTTP client to access the
  Serval DNA RESTful interface

Improve the REST /keyring/set operation to only alter the DID or Name if
the corresponding query parameter is supplied.  Modify the internal
keyring_set_did() function to only assign the DID or Name if the
corresponding parameter is not a null pointer.